+++ composeApp/src/commonMain/kotlin/fr/vetbrain/vetnutri_mp/Components/RationItem.kt
@@ -38,7 +38,7 @@ fun RationItem(
         onEdit: () -> Unit,
         onDelete: () -> Unit,
         onDuplicate: (() -> Unit)? = null,
-        onEditCoef: ((Float) -> Unit)? = null,
+        onEditCoef: ((Double) -> Unit)? = null,
         modifier: Modifier = Modifier
 ) {
         Card(
