+++ composeApp/src/commonMain/kotlin/fr/vetbrain/vetnutri_mp/Data/CoefP.kt
@@ -1,13 +1,14 @@
 package fr.vetbrain.vetnutri_mp.Data
 
 import fr.vetbrain.vetnutri_mp.Utils.genUUID
-import kotlinx.serialization.Serializable
 import kotlin.uuid.*
+import kotlinx.serialization.Serializable
+
 @OptIn(ExperimentalUuidApi::class)
 @Serializable
 data class CoefP(
-    val uuid: String = genUUID(),
-    var description: String?,
-    var coef: Float?,
-    var groupUUID: Int?
-) 
+        val uuid: String = genUUID(),
+        var description: String?,
+        var coef: Double?,
+        var groupUUID: Int?
+)
