+++ composeApp/src/commonMain/kotlin/fr/vetbrain/vetnutri_mp/View/AnalNut/AlimentsRationSection.kt
@@ -197,7 +197,7 @@ fun SectionAlimentsRation(
                                                                                                                                                 (alimentRation
                                                                                                                                                                 .quantite *
                                                                                                                                                                 ratio)
-                                                                                                                                                        .toFloat()
+                                                                                                                                                        .toDouble()
                                                                                                                                 )
                                                                                                                 }
                                                                                                 coroutineScope
